requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.4.4
Werkzeug==2.3.7
bcrypt
requests==2.31.0
